+++ azure-storage-queue/azure/storage/queue/queueservice.py
@@ -178,6 +178,7 @@ class QueueService(StorageClient):
             self.authentication = _StorageSharedKeyAuthentication(
                 self.account_name,
                 self.account_key,
+                self.is_emulated
             )
         elif self.sas_token:
             self.authentication = _StorageSASAuthentication(self.sas_token)